constraints: Fixes reeval layer constraint for layer cons > layerId
The former method for checking whether a set layer constraint needs to be increased based on an already set layer constrain > the layer id of its node was incorrect in certain situations in which shifted nodes are present.
The new solution should work for all cases.
Additionally, our message objects are extended. They include now the new layerId / posId and the constraint values since these values can be different.