fixed klighd-cli build.
Use polyfills for various packages not automatically polyfilled anymore with webpack >=5.

As this and the last commit removed and re-added klighd-vscode possibly a lot of versions in the yarn.lock have been changed unintentionally and possibly should be revoked (or kept for some more version updates) when merging into main.